+++ CHANGELOG.md
@@ -1,6 +1,6 @@
  ## 0.3.0 (TBR)
 
-  +
+  + Support for retrieving documentation from [Sanny Builder Library](https://library.sannybuilder.com) \[[#55](https://github.com/thelink2012/vscode-gta3script/pull/55) — contributed by [nikita488](https://github.com/nikita488)\]
  
  ## 0.2.9 (September 15 2024)